refactored feature
